trying a new  file structure
