+++ admin/src/App.jsx
@@ -4,9 +4,9 @@ import { Admin, Resource, EditGuesser } from 'react-admin';
 import { PostList } from './posts';
 import { TeacherList } from "./teachers";
 import { SchoolList } from "./schools";
-import { InvoiceList } from "./ivoices";
-import { PostEdit, TeacherEdit, SchoolEdit, InvoiceEdit} from "./post-edit";
-import { PostCreate, TeacherCreate, SchoolCreate, InvoiceCreate} from "./post-create";
+import { InvoiceList, InvoiceEdit,InvoiceCreate } from "./ivoices";
+import { PostEdit, TeacherEdit, SchoolEdit} from "./post-edit";
+import { PostCreate, TeacherCreate, SchoolCreate} from "./post-create";
 import { UserList } from './users';
 import jsonServerProvider from 'ra-data-json-server';
 import PostIcon from '@mui/icons-material/Book';
