+++ admin/src/ivoices.jsx
@@ -1,4 +1,4 @@
-import { Datagrid, List, TextField, EditButton } from 'react-admin';
+import { Datagrid, List, TextField, EditButton, Edit, SimpleForm, TextInput, Create,  } from 'react-admin';
 
 // ?? Need filter for List to be able to serach teachers later on ?? (Different feature)
 export const InvoiceList = () => (
@@ -16,3 +16,33 @@ export const InvoiceList = () => (
         </Datagrid>
     </List>
 );
+
+export const InvoiceCreate = props => (
+    <Create {...props}>
+        <SimpleForm>
+            <TextInput disabled source="id" />
+            <TextInput source="school" />
+            <TextInput source="customHeadline" />
+            <TextInput source="yourReference" />
+            <TextInput source="ourReference" />
+            <TextInput source="invoiceDate" />
+            <TextInput source="dueDate" />
+            <TextInput source="invoiceItems" />
+        </SimpleForm>
+    </Create>
+)
+
+export const InvoiceEdit = () => (
+    <Edit title="Edit Invoice">
+        <SimpleForm>
+            <TextInput disabled source="id" />
+            <TextInput source="school" />
+            <TextInput source="customHeadline" />
+            <TextInput source="yourReference" />
+            <TextInput source="ourReference" />
+            <TextInput source="invoiceDate" />
+            <TextInput source="dueDate" />
+            <TextInput source="invoiceItems" />
+        </SimpleForm>
+    </Edit>
+)
